minor changes-3 (raknet folder)

+++ src/handler/raknet_packet_handler.rs
@@ -24,10 +24,10 @@ pub struct RakNetPacketHandler {
     pub client_guid: i64,
     pub game: GamePacket,
     pub frame_number_cache: FrameNumberCache,
-    pub last_received_packets: HashMap<i32, Frame>, // reliable_frame_index: Frame
+    pub last_received_packets: HashMap<u32, Frame>, // reliable_frame_index: Frame
     pub last_received_fragment_packets: HashMap<u16, HashMap<u32, Vec<u8>>>, // split_id: index => buffer
-    pub last_received_sequence_number: i32,
-    pub last_handled_reliable_frame_index: i32
+    pub last_received_sequence_number: i64, // i64, u32'yi kapsadığı için kullandım (-1)
+    pub last_handled_reliable_frame_index: i64
 }
 
 impl RakNetPacketHandler {
@@ -53,7 +53,6 @@ impl RakNetPacketHandler {
     }
 
     pub fn handle_packet(&mut self, should_stop: &mut bool, debug: bool, target_address: String, target_port: u16, packet_type: PacketType, stream: &mut Stream) -> Vec<u8> {
-
         let mut response_data = vec![];
 
         match packet_type {
@@ -61,7 +60,7 @@ impl RakNetPacketHandler {
                 let open_conn_reply1 = OpenConnReply1::decode(Vec::from(stream.get_buffer()));
                 if debug { open_conn_reply1.debug(); }
 
-                response_data = OpenConnReq2::new(MAGIC, address::new(4, target_address.to_string(), target_port), open_conn_reply1.cookie, false, open_conn_reply1.mtu, self.client_guid).encode();
+                response_data = OpenConnReq2::new(MAGIC, address::new(4, target_address.to_string(), target_port), open_conn_reply1.cookie, false, open_conn_reply1.mtu, self.client_guid as u64).encode();
 
                 //client.socket.send(&req2).expect("Open Connection Request 2 Packet could not be sent");
             },
@@ -69,7 +68,7 @@ impl RakNetPacketHandler {
                 let open_conn_reply2 = OpenConnReply2::decode(Vec::from(stream.get_buffer()));
                 if debug { open_conn_reply2.debug(); }
 
-                let body = ConnReq::new(self.client_guid, Utc::now().timestamp(), false).encode();
+                let body = ConnReq::new(self.client_guid as u64, Utc::now().timestamp() as u64, false).encode();
 
                 let frame = Datagram::create_frame(body, RELIABLE, &self.frame_number_cache, None);
                 response_data = Datagram::create(vec![frame], &self.frame_number_cache).to_binary();
@@ -84,13 +83,13 @@ impl RakNetPacketHandler {
 
                 // New Incoming Connection
                 let addresses: [InternetAddress; 20] = core::array::from_fn(|_| address::new(4, "0.0.0.0".to_string(), 0));
-                let new_incoming_conn = NewIncomingConn::new(address::new(4, target_address.to_string(), target_port), addresses, Utc::now().timestamp(), Utc::now().timestamp() + 1).encode();
+                let new_incoming_conn = NewIncomingConn::new(address::new(4, target_address.to_string(), target_port), addresses, Utc::now().timestamp() as u64, (Utc::now().timestamp() + 1) as u64).encode();
                 let frame = Datagram::create_frame(new_incoming_conn, RELIABLE_ORDERED, &self.frame_number_cache, None);
                 self.frame_number_cache.reliable_frame_index += 1;
                 self.frame_number_cache.ordered_frame_index += 1;
 
                 // Connected Ping
-                let connected_ping = ConnectedPing::create(Utc::now().timestamp()).encode();
+                let connected_ping = ConnectedPing::create(Utc::now().timestamp() as u64).encode();
                 let frame_two = Datagram::create_frame(connected_ping, UNRELIABLE, &self.frame_number_cache, None);
 
                 // Request Network Settings Packet
